fix: UserAtom 여러번 정의되는 문제 수정
